+++ returns/future.py
@@ -28,6 +28,7 @@ from returns.primitives.hkt import (
     SupportsKind2,
     dekind,
 )
+from returns.primitives.iterables import BaseIterableStrategyN, FailFast
 from returns.result import Failure, Result, Success
 
 # Definitions:
@@ -381,7 +382,9 @@ class Future(
 
     @classmethod
     def from_iterable(
-        cls, inner_value: Iterable[Kind1['Future', _NewValueType]],
+        cls,
+        inner_value: Iterable[Kind1['Future', _NewValueType]],
+        strategy: Type[BaseIterableStrategyN] = FailFast,
     ) -> 'Future[Sequence[_NewValueType]]':
         """
         Transforms an iterable of ``Future`` containers into a single container.
@@ -398,7 +401,7 @@ class Future(
           ... ]).awaitable) == IO((1, 2))
 
         """
-        return dekind(iterable_kind(cls, inner_value))
+        return iterable_kind(cls, inner_value, strategy)
 
     @classmethod
     def from_io(cls, inner_value: IO[_NewValueType]) -> 'Future[_NewValueType]':
@@ -1378,6 +1381,7 @@ class FutureResult(
         inner_value: Iterable[
             Kind2['FutureResult', _NewValueType, _NewErrorType],
         ],
+        strategy: Type[BaseIterableStrategyN] = FailFast,
     ) -> 'FutureResult[Sequence[_NewValueType], _NewErrorType]':
         """
         Transforms an iterable of ``FutureResult`` containers.
@@ -1388,7 +1392,7 @@ class FutureResult(
 
           >>> import anyio
           >>> from returns.future import FutureResult
-          >>> from returns.io import IOSuccess, IOFailure
+          >>> from returns.io import IOSuccess
 
           >>> all_success = FutureResult.from_iterable([
           ...    FutureResult.from_value(1),
@@ -1396,20 +1400,9 @@ class FutureResult(
           ... ])
           >>> assert anyio.run(all_success.awaitable) == IOSuccess((1, 2))
 
-          >>> mixed = FutureResult.from_iterable([
-          ...     FutureResult.from_value(1),
-          ...     FutureResult.from_failure('a'),
-          ... ])
-          >>> assert anyio.run(mixed.awaitable) == IOFailure('a')
-
-          >>> mixed = FutureResult.from_iterable([
-          ...     FutureResult.from_failure('a'),
-          ...     FutureResult.from_value(1),
-          ... ])
-          >>> assert anyio.run(mixed.awaitable) == IOFailure('a')
 
         """
-        return dekind(iterable_kind(cls, inner_value))
+        return iterable_kind(cls, inner_value, strategy)
 
 
 def FutureSuccess(  # noqa: N802