+++ returns/maybe.py
@@ -16,11 +16,12 @@ from typing import (
 from typing_extensions import final
 
 from returns._generated.iterable import iterable_kind
-from returns.interfaces import iterable, unwrappable
+from returns.interfaces import rescuable, unwrappable
 from returns.interfaces.aliases.container import Container1
 from returns.primitives.container import BaseContainer
 from returns.primitives.exceptions import UnwrapFailedError
-from returns.primitives.hkt import Kind1, SupportsKind1, dekind
+from returns.primitives.hkt import Kind1, SupportsKind1
+from returns.primitives.iterables import BaseIterableStrategyN, FailFast
 
 # Definitions:
 _ValueType = TypeVar('_ValueType', covariant=True)
@@ -35,8 +36,8 @@ class Maybe(
     BaseContainer,
     SupportsKind1['Maybe', _ValueType],
     Container1[_ValueType],
+    rescuable.Rescuable2[_ValueType, None],
     unwrappable.Unwrappable[_ValueType, None],
-    iterable.Iterable1[_ValueType],
     metaclass=ABCMeta,
 ):
     """
@@ -117,6 +118,28 @@ class Maybe(
 
         """
 
+    def rescue(
+        self,
+        function: Callable[[None], Kind1['Maybe', _ValueType]],
+    ) -> 'Maybe[_ValueType]':
+        """
+        Composes failed container with a function that returns a container.
+
+        .. code:: python
+
+          >>> from returns.maybe import Maybe, Some, Nothing
+
+          >>> def rescuable(arg=None) -> Maybe[str]:
+          ...      return Some('b')
+
+          >>> assert Some('a').rescue(rescuable) == Some('a')
+          >>> assert Nothing.rescue(rescuable) == Some('b')
+
+        We need this feature to make ``Maybe`` compatible
+        with different ``Result`` like oeprations.
+
+        """
+
     def value_or(
         self,
         default_value: _NewValueType,
@@ -141,8 +164,10 @@ class Maybe(
 
         Really close to :meth:`~Maybe.value_or` but works with lazy values.
         This method is unique to ``Maybe`` container, because other containers
-        do have ``.rescue``, ``.alt``, ``.fix`` methods.
-        But, ``Maybe`` does not.
+        do have ``.alt`` method.
+
+        But, ``Maybe`` does not have this method.
+        There's nothing to ``alt`` in ``Nothing``.
 
         Instead, it has this method to execute
         some function if called on a failed container:
@@ -221,31 +246,22 @@ class Maybe(
     def from_iterable(
         cls,
         inner_value: Iterable[Kind1['Maybe', _NewValueType]],
+        strategy: Type[BaseIterableStrategyN] = FailFast,
     ) -> 'Maybe[Sequence[_NewValueType]]':
         """
         Transforms an iterable of ``Maybe`` containers into a single container.
 
         .. code:: python
 
-          >>> from returns.maybe import Maybe, Some, Nothing
+          >>> from returns.maybe import Maybe, Some
 
           >>> assert Maybe.from_iterable([
           ...    Some(1),
           ...    Some(2),
           ... ]) == Some((1, 2))
 
-          >>> assert Maybe.from_iterable([
-          ...     Some(1),
-          ...     Nothing,
-          ... ]) == Nothing
-
-          >>> assert Maybe.from_iterable([
-          ...     Nothing,
-          ...     Some(1),
-          ... ]) == Nothing
-
         """
-        return dekind(iterable_kind(cls, inner_value))
+        return iterable_kind(cls, inner_value, strategy)
 
 
 @final
@@ -280,6 +296,10 @@ class _Nothing(Maybe[Any]):
     def bind(self, function):
         """Does nothing for ``Nothing``."""
         return self
+
+    def rescue(self, function):
+        """Composes this container with a function returning container."""
+        return function(None)
 
     def value_or(self, default_value):
         """Returns default value."""
@@ -330,6 +350,10 @@ class _Some(Maybe[_ValueType]):
     def bind(self, function):
         """Binds current container to a function that returns container."""
         return function(self._inner_value)
+
+    def rescue(self, function):
+        """Does nothing for ``Some``."""
+        return self
 
     def value_or(self, default_value):
         """Returns inner value for successful container."""