+++ CHANGELOG.md
@@ -44,11 +44,10 @@ See [0Ver](https://0ver.org/).
   to be similar to `Result` and `IOResult`
 - Adds `.swap` method to `Result`, `IOResult`, `FutureResult`,
   and other result based containers
+- Adds `.modify_env` method to all `RequiresContext*` types
+- Adds `.rescue` to `Maybe`
 - Adds `collect_trace` helper function for better development experience
 - Adds missing `from_requires_context_future_result` to `RequiresContext`
-- Adds `.modify_env` method to all `RequiresContext*` types
-- Adds `@abstractmethod` to abstract methods in abstract containers
-  like `Result`, `Maybe`, and `IOResult`
 - Adds `__slots__` to `UnwrapFailedError` with `halted_container`
 - Changes `flatten` to work with `KindN` and any possible container
 - Changes `iterable` helper function to be `iterable_kind`,